requirements.txt
backpack_for_pytorch==1.5.0
numpy==1.16.4
Pillow==6.0.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.32.1
